Forcibly set a Dialtimeout for promhttputil.Client
If a node is in the sd response but unavailable there is currently
nothing to time out the request (barring OS timeouts-- which are long).
Ideally we'd be able to (1) set a dial timeout in config or (2)
configure sd to exclude "unhealthy" hosts. Since neither of those exists
for now we'll fore a 200ms dial timeout.