Change QueryRange to switch between raw and range data queries
Although changing all the queries to be a simple range fetch
(metric[range]) gives us more granular results, it turns out that this
is a HUGE burden on prometheus (specifically because its very
inefficient at json marshalling). This is being put in here as a
hard-coded stop-gap. Ideally this would be configuration controlled so
you could decide at what point you'd like to drop granularity in favor
of load reduction