Implement NodeReplacer in proxystorage
This allows promxy to break the incoming promql statement into
sub-queries that can be farmed out to the various nodes in the
servergroups. This significantly reduces the amount of data transfer
across the wire. As such performance with this is roughly on-par with
going direct to prometheus (with the usual caveat that it is bounded by
the slowest one)