Stop proxying prometheus UI pages
Promxy until now has been proxying everything not under /api to a random server downstream. This works for the /graph pages as they simply call the /api endpoints. As promxy supports alerts, reloading, etc. it is becoming increasingly obvious that proxying those endpoints is becoming confusing.

This adds the "normal" prometheus web UI to promxy. The only caveats are:

- scrape targets will always be empty: promxy right now doesn't do any scraping
- config is only prometheus config: prometheus' config endpoint doesn't take the promxy struct, as such it only shows the prometheus configs

This fixes #28