Ensure consistent return order when querying a servergroup
Without this whichever target responded first was the "base". This works
for most things, but when there are rates etc. it can sometimes differ.
A noticeable case was doing a query like sum(subset) / sum(totalset) and
you would sometimes get values >1 -- this was due to the fact that the
numbers in the query weren't coming from the same server in the
servergroup. This patch makes a best-effort attempt to do so. There is
still a possibility for an issue if the servergroup target list changes
mid-query, but otherwise its resolved