Only set dialtimeout for http.Transport
To workaround the lack of health-checking in the prom SD libraries
promxy was setting a dial timeout. If promxy was configued with basic
auth (or any other auth for that matter) the roundtripper wasn't the
default one and this would panic. This changes it to only apply when the
type matches, and is left type-switched in case upstream changes their
structs to expose those.

Fixes #70