Add support for relabel_config to add labels in a servergroup
Before the relabel_config was used soely to remove hosts from the server
group (to have more specific targeting). This change allows the user to
define a regular relabel_config which will be used to add those labels
to all series returned by the servergroup. One thing to keep in mind is
that each host in the servergroup should return the same labels, this is
enforced in the code and will log a warning when it has to step in to
make them the same.