Move servergroup label filtering
Now that we have nice layer-able promclients it makes more sense to
consolidate this logic into the promclient instead of into the
servergroup.

In addition to the simplification in code, this has the added benefit of
allowing the user to have a lot more control over the merge behavior.
Before this patch if there was a difference in value between the
relabel_config output it would be ignored and a warning would be
printed. Now if there is a difference it will still be applied-- which
means you can effectively control merging behavior.