Add better controls for error-handling in multi_api client
Before this change the multi_api client would return a non-error respons
if a single `api` it had returned a non-error response. This has a few
problems (1) if there are multiple "types" of data -- based on
labels/relabel_config (2) global level -- e.g. cross servergroup
merging.

In (1) if we had N shards of data, we would only require 1 host (not
shard) to return a non-error. This means we could be knowingly missing
data but still returning a non-error response

In (2) if any servergroup returned a non-error, the response will be
non-error.

This patch adds a `requiredCount` option to MultiAPI which is the number
of required responses for a given key.