Enable graceful shutdown behavior
This (1) overrides the ready check, and we start failing it as soon as we get a stop signal (2) this adds an option to sleep between stopping everything and calling srv.Shutdown.

These can be combined to create a "0 downtime" scale up/down. This can be done by setting ShutdownDelay to >= the loadbalancer healthcheck interval, and setting ShutdownTimeout to whatever timeout you are willing to tolerate past that.

Fixes #143