Add NodeReplacer for VectorSelector
If we are simply selecting a vector we don't need to use the raw query
selection -- as we are getting specific points in time (not raw data).
The comment here made it sound like it was *already* doing this but in
fact it wasn't. So this should be a significant performance improvement
for queries that the base is a vectorselector.

Fixes #188