Use DisableLookback feature in NodeReplacer
In NodeReplacer we are sub-tasking the queries to remote prom APIs; if
we do this we should not do our own LookbackDelta on top of it -- as it
was already done remotely.

Prior to this if the remote returned a result with gaps we'd fill them
as long as there was a point within LookbackDelta.