Re-enable upstream tests
I found this because of the last 2 bugs reported to promxy -- which should have been caught by these tests. These were removed in some previous release due to prune options. Unfortunately this means we end up having to vendor in a lot of unused code from the prometheus repo -- but we do get the upstream tests.