Add support for promql Subquerys
It turns out adding support for Subqueries wasn't as simple as upgrading
the prom dep to 2.10+. Subqueries are an intersting addition to promql
which effectively let you select a matrix anywhere in the query.
Unfortunately this is special-cased heavily in promql so to make this
work in NodeReplacer we need to add/use a RawMatrix type which the
promql engine will just use as a return type instead of trying to select
data within the matrix.

Fixes #245