Properly re-run aggregates for subqueries
Previously we were replacing the returned Expr from an AggregateExpr to
a MatrixSelector -- meaning it wasn't aggregated. This was causing some
interesting promql failures (as a `sum` would have a bunch of labels --
even though there was no `by` clause). This simpy wraps the resultant
Selector within the AggregateExpr to be a Raw selector and still keep
the aggregateExpr wrapping.

Fixes #273