Revert "Use DisableLookback feature in NodeReplacer"
This reverts commit c2454bb3d4081cec103a06c99eedaa67d1d140fe.

in #223 the objective was to avoid querying excessive amounts of
previous data (e.g. double lookback range). This particular change
actually made it so we are now significantly more strict about the
timestamps that we get from the downstream; which is not as desierable
(as any level of caching in the middle causes issues). What this does
mean though is if the result returns an inaccurate timestamp (due to
caching or something else) promql may interpolate the value -- which is
what it does for all other queries of data off of the scrape interval.
So this should be fine -- and was the previous behavior.

Fixes #298