In merging start with the series with the most data
The majority of times there are holes in the data a single downstream
has a hole but the other has the data; in that case since we have the
data in memory there is no reason to chose the "worse" data and merge
from there.

This has the caveat that this is done on a per-merge basis; so if there
are N servers and the first 2 return with holes they will be merged; but
due to anti-affinity if there is any server with no hole it will always
have more points than a merged series.