Set RemoteReadConcurrencyLimit to fix /api/v1/read
Previously this value was unset, defaulting to zero. A zero-value
initialised a channel with no buffer, causing all writes to the buffer
to hang, and therefore all remote_read queries to hang until they
timed-out. This is contradictory to the Prometheus documentation, which
claims that a zero value allows infinite concurrent reads.

Fixes: https://github.com/jacksontj/promxy/issues/350
Signed-off-by: Joe Groocock <jgroocock@cloudflare.com>