Fix subquery handling
Now NodeReplacer treats subquery as effectively a new EvalStmt. This
dramatically simplifies the codebase (as we don't have to handle all the
various edge cases) and allows us to generically support Subquerys. The
only catch here is that the data passed down will be interpreted with
the step requested in the query (which is what the spec calls for)
unlike prom which actually fetches all raw data when using its own tsdb
direct. Because of this there are some slight deltas based on the
interpolation of step values in promql but this is unfortunately
unavoidable.

Fixes #273