Fix MatrixSelector downstream querying
In the new parser format (included in the prom fork upgrade) the way
MatrixSelectors worked changed. Unfortunately this broke functionality
as we ended up treating them as VectorSelectors (since MatrixSelectors
now include VectorSelectors -- so the tree replacement handled it). This
changes the VectorSelector in the NodeReplacer to ignore VectorSelectors
that are children of a MatrixSelector.

Fixes #409