Correct series Select handling with upstream changes
In prometheus 2.24 dep upgrade prometheus changed how it signaled to
queriers whether the request was for data or just the series.
So with that upgrade we started *always* fetching all the data
regardless of the query (in this case a series call would fetch all data
for those series in the time range requested). As with before we're
working around this by introspecting (cortex does the same --
https://github.com/cortexproject/cortex/pull/3461).

Fixes #415