Add DropStale flag to remote_write_exporter
DropStale will remove series from the map when a StaleNan is sent
(instead of proxying that staleNan value downstream). This is important
as the text scrapers only get a NaN and can't distinguish between a
StaleNaN and any other Nan. Leaving this as default false for now as a
backwards compatiblity, but likely this will change in the future.

Fixes #440