Add metric relabel support to promxy
This is a fairly large feature, but the intent is to allow some subset
of relabel capability to the metrics that flow through promxy. This is
pretty heavily restricted because the rewrites need to be bi-directional
(as we need to be able to modify the query we are sending downstream).

This enables a few major use-cases:
  - dropping labels (e.g. thanos replica labels)
  - renaming labels (e.g. migrating to a new label name)

Fixes #258