Fix race condition in CreateAndStart
Today the CreateAndStart method starts the listener in a goroutine
async. This is not a problem for regular operations, but during testing
this creates a race condition. This changes the behavior so that listen
will for-sure be started when the method returns and the hanlder will be
in a background goroutine.