Correct OffsetFinder to handle "new" SubQuery support
Offsetfinder is responsible to detemrine that the "tree" we are working
on is at the same time offset. With the addition of SubQuery you can
actually have an entire sub-tree at its own offset. Since subquery
support was added (#246) we can actually allow NodeReplacer to run
assuming the top of the tree is a SubQuery -- as we will "expand" it
into its own tree.