Correctly override node in aggregateBinaryExpr
In #676 we added support for query push down for BinaryExp where it is a
literal and either a VectorSelector or a subset of AggregateExprs.
However this initial implementation introduced a bug for AggregateExprs.
Specifically it was doing the query push down but not doing the node
replace properly -- such that the literal operation was done twice (e.g.
add twice, multiply twice, etc.).

This patch will have the aggregateBinaryExpr replace the Binary Expr
(i.e. `(max(foo) * 100)` ) with the "filled in" AggregateExpr -- such
that the re-execution of the aggregation still occurs without re-doing
the literal calculation.

Fixes #688