Update base image
The hardcoded 3.17.2 version of alpine used to build the final image has known vulnerabilities which lead to promxy being flagged by scanners (even though it's likely that promxy installations aren't actually affected). 

Bumping the version to the latest patch of the 3.18 series aligns it with what's used in the Go build, but both are not the latest available – Go 1.23 and Alpine 3.21 are available.