Allow for NodeReplacer on SubQuery expressions with Aggregates
Previously we would disallow any NodeReplacement for statements with
more than 1 aggregation in them. This logic was in-place well before
SubQuery support; and now needs some adjustment. In all cases its safe
to let SubQuery run as it recursively calls NodeReplacer with its
subquery.

For an example query of `avg_over_time(sum(count(up))[10m:15s])` we
would previously do a `GetValue` of `up`.

With this change we now do a `queryRange` for `count(up)`

Fixes #723